requirements.txt
affine==2.4.0
attrs==23.1.0
beautifulsoup4==4.12.2
blessings==1.7
blinker==1.7.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
dnspython==2.4.2
docopt==0.6.2
earthengine-api==0.1.371
ee==0.2
exceptiongroup==1.1.3
ffmpeg-python==0.2.0
filelock==3.12.2
fiona==1.9.5
Flask==3.0.1
fonttools==4.42.1
fsspec==2023.6.0
future==0.18.3
geopandas==0.14.2
google-api-core==2.12.0
google-api-python-client==2.101.0
google-auth==2.23.2
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
h11==0.14.0
httplib2==0.22.0
huggingface-hub==0.16.4
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
llvmlite==0.40.1
MarkupSafe==2.1.3
matplotlib==3.7.2
more-itertools==10.1.0
MouseInfo==0.1.3
mpmath==1.3.0
networkx==3.1
numba==0.57.1
numpy==1.24.4
openai-whisper==20230124
opencv-python==4.8.0.76
outcome==1.2.0
packaging==23.1
pandas==2.2.0
Pillow==10.0.0
pipreqs==0.4.13
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
pymongo==4.4.1
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
pyproj==3.6.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
pytube==15.0.0
pytweening==1.0.7
pytz==2023.3.post1
PyYAML==6.0.1
rasterio==1.3.9
regex==2023.8.8
requests==2.31.0
rsa==4.9
safetensors==0.3.2
selenium==4.8.0
semantic-version==2.10.0
setuptools-rust==1.5.2
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
snuggs==1.4.7
sortedcontainers==2.4.0
soupsieve==2.4.1
sympy==1.12
tiktoken==0.3.1
tokenizers==0.13.2
torch==2.0.1
tqdm==4.66.1
transformers==4.31.0
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.7.1
tzdata==2023.4
uritemplate==4.1.1
urllib3==1.26.16
Wave==0.0.2
Werkzeug==3.0.1
wsproto==1.2.0
yarg==0.1.9
